# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI短片工坊",
    "description": "本地单用户短片工坊：M1故事版本与DeepSeek适配器；真实文本验收及后续成片能力单独验证。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/settings/asset-credentials": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Asset Credentials",
        "operationId": "get_asset_credentials_api_v1_settings_asset_credentials_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetCredentialStatus"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Put Asset Credentials",
        "operationId": "put_asset_credentials_api_v1_settings_asset_credentials_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCredentialSave"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetCredentialStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{pid}/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Submit",
        "operationId": "submit_api_v1_projects__pid__jobs_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pid"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_v1_projects__pid__jobs_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobOut"
                  },
                  "title": "Response List Jobs Api V1 Projects  Pid  Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jid}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_jobs__jid__get",
        "parameters": [
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Jid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetConfiguration": {
        "properties": {
          "bucket": {
            "type": "string",
            "maxLength": 63,
            "minLength": 3,
            "pattern": "^[a-z0-9][a-z0-9-]*[a-z0-9]$",
            "title": "Bucket"
          },
          "region": {
            "type": "string",
            "const": "cn-beijing",
            "title": "Region",
            "default": "cn-beijing"
          },
          "project_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Project Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bucket",
          "project_name"
        ],
        "title": "AssetConfiguration"
      },
      "AssetCredentialSave": {
        "properties": {
          "bucket": {
            "type": "string",
            "maxLength": 63,
            "minLength": 3,
            "pattern": "^[a-z0-9][a-z0-9-]*[a-z0-9]$",
            "title": "Bucket"
          },
          "region": {
            "type": "string",
            "const": "cn-beijing",
            "title": "Region",
            "default": "cn-beijing"
          },
          "project_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Project Name"
          },
          "base_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Base Version"
          },
          "access_key": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "format": "password",
            "title": "Access Key",
            "writeOnly": true
          },
          "secret_key": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "format": "password",
            "title": "Secret Key",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bucket",
          "project_name",
          "base_version",
          "access_key",
          "secret_key"
        ],
        "title": "AssetCredentialSave"
      },
      "AssetCredentialStatus": {
        "properties": {
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "configuration": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetConfiguration"
              },
              {
                "type": "null"
              }
            ]
          },
          "connection_state": {
            "type": "string",
            "const": "not_tested",
            "title": "Connection State",
            "default": "not_tested"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "configured",
          "revision"
        ],
        "title": "AssetCredentialStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCreate": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "file.verify",
            "title": "Kind"
          },
          "file_id": {
            "type": "string",
            "format": "uuid",
            "title": "File Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "file_id"
        ],
        "title": "JobCreate"
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Snapshot"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "kind",
          "state",
          "snapshot",
          "error",
          "created_at",
          "updated_at"
        ],
        "title": "JobOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}